requirements.txt
annotated-types==0.7.0
certifi==2026.1.4
charset-normalizer==3.4.4
dnspython==2.7.0
email-validator==2.3.0
exceptiongroup==1.3.1
Faker==37.12.0
idna==3.11
iniconfig==2.1.0
Jinja2==3.1.6
MarkupSafe==3.0.3
packaging==26.0
pluggy==1.6.0
pydantic==2.12.5
pydantic_core==2.41.5
Pygments==2.19.2
PyJWT==2.11.0
pytest==8.4.2
pytest-html==4.2.0
pytest-metadata==3.1.1
python-dotenv==1.2.1
requests==2.32.5
requests-mock==1.12.1
tomli==2.4.0
typing-inspection==0.4.2
typing_extensions==4.15.0
tzdata==2025.3
urllib3==2.6.3
